requirements.txt
aiofiles==24.1.0
aiohttp==3.8.1
cachetools==5.5.2
certifi==2022.6.15
faiss_cpu==1.10.0
fastapi==0.115.12
httpx==0.23.0
loguru==0.6.0
matplotlib==3.10.1
nltk==3.9.1
nonebot-adapter-qq==1.6.2
nonebot-plugin-alconna==0.57.2
nonebot-plugin-waiter==0.8.1
nonebot2==2.4.2
numpy==2.2.4
opencc_python_reimplemented==0.1.7
Pillow
pycryptodome==3.21.0
pydantic==1.10.19
PyQt6==6.9.0
python_Levenshtein==0.27.1
python_rapidjson==1.20
pytz==2025.1
Requests==2.32.3
scikit_learn==1.6.1
selenium==4.31.0
selenium_wire==5.1.0
sentence_transformers==3.4.1
torch==2.6.0
torchvision==0.21.0
uvicorn==0.34.0
wget==3.2
blinker==1.7.0
